fixed undo redo for synthesis
